1750: Add CDROM configuration to "Create VM" wizard:
First Run Wizard implemented.
This dialog allows user to select and temporarily mount the booting source for the VM to be started for the first time. The booting source could be:
1. CD/DVD-ROM host drive.
2. CD/DVD-ROM image file.
3. Floppy host drive.
4. Floppy image file.
It also allows user to invoke Virtual Disk Manager Dialog to add additional CD/FD images as booting sources.

The dialog will only be executed for the VM which has a GUI/FirstRun=”yes” extra data flag in it's configuration file. This flag currently will be auto-added only for the new VM created by the “New VM Wizard”. After the “First Run Wizard” will be proposed for the first time, the flag will be erased from the VM configuration file. This flag will also be erased from the configuration file in case of there is any other value rather than “yes” for it.

TODO: The flag should also be erased from the configuration file in case of user manually changes the boot sequence or the CD/FD/HD configuration through the VM Settings Dialog.

git-svn-id: https://www.virtualbox.org/svn/vbox@2864 cfe28804-0f27-0410-a406-dd0f0b0b656f