Dual cursor issue fixed.
Issue description: If the GuestOS is loaded (in case of "Guest Additions" are correctly installed and activated and "Mouse integration" feature is activated too) and the mouse cursor was not moved during the guest loading process then the user can see two mouse cursors - guest's and host's, which are positioned separately. If the user moves the mouse cursor above the VM Console Viewport, both cursors will be positioned one-above-one. It looks like some dirty visual artifact.

git-svn-id: https://www.virtualbox.org/svn/vbox@3063 cfe28804-0f27-0410-a406-dd0f0b0b656f