NAT: Rework sorecvfrom().
On Windows provide iovec definition that is congruent to WSABUF (and
assert at compile time that it is).  Get rid of FIONREAD hack, instead
use recvmsg()/WSARecvFrom to read the datagram into the first mbuf and
a spillover buffer large enough to accomodate anything.  Small
datagrams will fit and will need to copying, for large datagrams
m_append() spilover buffer contents.

Fix mbuf length calculations - old code would cause memory corruption
for certain datagram sizes that it considered small enough when they
weren't.

Untanlge the code a bit and make sure we do SOCKET_UNLOCK() on error
legs too (though it happens to be a nop, so we were lucky).


git-svn-id: https://www.virtualbox.org/svn/vbox@52154 cfe28804-0f27-0410-a406-dd0f0b0b656f