bugref:8533: Additions/x11: fully support VMSVGA Allow VBoxClient services to daemonise without providing a pidfile.
Up until now VBoxClient service processes would be made into daemons (unless
the user overrid this) if they provided a name for creating a pidfile.  We
add an additional member to the service description structure which is a
boolean specifying whether the service should be daemonised anyway.  The
vmsvga service does not need a pidfile as it will use guest capability
acquiring to prevent multiple copies running, but should still be a daemon.
As the new structure member is at the end, it will automatically be zero
(false) for existing services which do not set it.

git-svn-id: https://www.virtualbox.org/svn/vbox@67319 cfe28804-0f27-0410-a406-dd0f0b0b656f