Linux driver Make files: fix Linux 2.4 module installation. bugref:4567: Linux kernel driver maintenance
The Guest Additions installer was no longer installing kernel modules on
Linux 2.4-based systems due to a path specification problem which was a
regression from recent clean-ups.  Fix this and switch to taking a kernel
version rather than a kernel header path as a parameter for allowing builds
for different kernels.

git-svn-id: https://www.virtualbox.org/svn/vbox@68706 cfe28804-0f27-0410-a406-dd0f0b0b656f