Additions/x11/vboxvideo: always pull in stdint.h, even on XFree86. bugref:9017: Additions/x11: put vboxvideo into upstream X.Org
Because XFree86 tried to provide its own versions of everything, including
part of the standard C library, including standard headers when building
XFree86 libraries tends to cause problems, and I was trying to recreate
part of stdint.h in the IPRT replacement header file.  Reading the XFree86
header files and testing seem to indicate though that including stdint.h is
doable.  So this change removes the stdint.h replacement code, which was
causing build problems on Debian 9.0, and includes stdint.h unconditionally.

git-svn-id: https://www.virtualbox.org/svn/vbox@69085 cfe28804-0f27-0410-a406-dd0f0b0b656f