Additions/WINNT/SharedFolders: Provide a dedicated 32bit VBoxSF driver for Windows <= 7 linked against the W2K3 DDK rdbss.lib and link the default one against the WHL version for Windows >= 8
The reason for shipping two binaries now is that the mmap FsPerf tests failed under certain circumstances, FlushViewOfFile() would return ERROR_INVALID_PARAMETER.
This only happens when the file is opened readonly first, and then opened read/write again while the first open is still active. The test would go on creating
a mapping of the file using the read/write handle, close both handles and then modify the mapping. Upon flushing the mapping ERROR_INVALID_PARAMETER is returned
because RDBSS passes the readonly file handle to our shared folder driver which the host shared folder service complains about rightfully because it is readonly.
The request returns VERR_INVALID_HANDLE for the write which gets translated to ERROR_INVALID_PARAMETER.
This only happens on 32bit Windows 8 and newer when linked against the W2K3 rdbss.lib, the reason is unknown.
Using the WHL version solves this and to not run into weird compatibility issues on older versions we ship two variants now and the installer decides during
install time which one is used.



git-svn-id: https://www.virtualbox.org/svn/vbox@84972 cfe28804-0f27-0410-a406-dd0f0b0b656f